NetflixStockPriceAnalysis
Analysis of the stock price, and determine the volume of stock traded for Netflix. Identify the high, open, and close prices of Netflix stock on a daily, monthly, and yearly basis. Provide the top 5 dates with the highest stock prices and the top 5 dates with the lowest stock prices in this project. Utilize NumPy and Pandas for data cleaning, data preprocessing, and Matplotlib for visualization.